Fix spreadsheet initialization bug

+++ build-elastic-sheets-project.sh
@@ -1,6 +1,15 @@
 #!/bin/bash
-cd elastic-sheets-project
+if [ $# -ne 1 ]; then
+   echo "build-elastic-sheets-project <dir>"
+   exit -1
+fi
+if [ ! -d "$1" ]; then
+   echo "dir ./$1 does not exist"
+   exit -1
+fi
+cd $1 
 #clasp create elasticSheetsProjects
+#clasp create elasticSheetsShare
 rm -f *.js *.gs *.html
 for i in $(find ../src ../test/server -name "*.gs" -o -name "*.html"); do cp $i .; done
 for i in $(find ../src ../test/server -name "*.js"); do
